feat: LoggedInWrapper util로 분리
