+++ frontend/src/components/Profile/Profile.stories.tsx
@@ -1,29 +1,11 @@
 import { Story } from "@storybook/react";
-import { useContext, useEffect } from "react";
-import UserContext, { UserContextProvider } from "../../contexts/UserContext";
 
+import LoggedInWrapper from "../../../.storybook/utils/LoggedInWrapper";
 import Profile, { Props } from "./Profile";
 
 export default {
   title: "Components/Profile",
   component: Profile,
-};
-
-const LoggedInWrapper = ({ children }: { children: React.ReactElement }) => {
-  const Inner = () => {
-    const { login } = useContext(UserContext);
-
-    useEffect(() => login("test", "Tanney"), []);
-
-    return <></>;
-  };
-
-  return (
-    <UserContextProvider>
-      <Inner />
-      {children}
-    </UserContextProvider>
-  );
 };
 
 const Template: Story<Props> = (args) => (
